backend/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt>=4.1.0
python-jose[cryptography]
python-multipart
pillow
numpy
opencv-python-headless
tensorflow
torch
torchvision
mtcnn; python_version < "3.13"
